+++ docs/changelog.md
@@ -8,6 +8,7 @@ nav_order: 99
 ### vNext
 - Mapping now checks if the root object is contained in the registered Mappers
 - Add a simple string as line in the mapper
+- Show more information about the missmatch in the exception
 
 ### 2.0.2
 - Mocking Guids also adds a directive to replace string Guids